requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
greenlet==3.0.3
idna==3.7
requests==2.32.3
SQLAlchemy==2.0.31
typing_extensions==4.12.2
urllib3==2.2.2

python-dotenv~=1.0.1
alembic~=1.13.2
pytest~=8.3.1